problem with Null/Null

+++ dzAlerts/util/env/elasticsearch.py
@@ -299,7 +299,7 @@ class ElasticSearch(object):
                     })
 
             if self.debug:
-                Log.note("{{num}} items added", {"num": len(lines) / 2})
+                Log.note("{{num}} items added", {"num": int(len(lines) / 2)})
         except Exception, e:
             if e.message.startswith("sequence item "):
                 Log.error("problem with {{data}}", {"data": repr(lines[int(e.message[14:16].strip())])}, e)
